Utilize shared utility functions and improve code organization
