+++ resources/js/Components/Public/Search/SearchResultSection.vue
@@ -53,14 +53,8 @@ import { computed, ref, watch, inject } from 'vue'
 import { AisHighlight } from 'vue-instantsearch/vue3/es'
 import { Button } from '@/Components/ui/button'
 import { trans as $t } from 'laravel-vue-i18n'
-import { format } from 'date-fns'
 import { useSearchService, type SearchService } from '@/Composables/useSearchService'
-
-// Import icons using the ~icons pattern like DutyForm.vue
-import IconNews from '~icons/fluent/news20-regular'
-import IconPage from '~icons/fluent/document-text20-regular'
-import IconDocument from '~icons/fluent/document20-regular'
-import IconCalendar from '~icons/fluent/calendar20-regular'
+import { useSearchUtils } from '@/Composables/useSearchUtils'
 import IconArrowRight from '~icons/fluent/arrow-right-16-filled'
 
 interface Props {
@@ -83,6 +77,19 @@ const emit = defineEmits<{
 
 // Use centralized search service - inject from parent or create new
 const searchService = inject<SearchService>('searchService') || useSearchService()
+
+// Use shared search utilities
+const { 
+  getIconComponent, 
+  formatDate, 
+  stripHtml, 
+  getItemDate, 
+  getItemContent,
+  getSectionHeaderClasses,
+  getItemClasses,
+  getIconClasses,
+  getLoadMoreClasses
+} = useSearchUtils()
 
 // Get results for this content type from the search service
 const typeResults = computed(() => {
@@ -124,94 +131,5 @@ watch(typeResults, (newTypeResults) => {
   }
 }, { immediate: true })
 
-const getIconComponent = (type: string) => {
-  switch (type) {
-    case 'news': return IconNews
-    case 'pages': return IconPage
-    case 'documents': return IconDocument
-    case 'calendar': return IconCalendar
-    default: return IconPage
-  }
-}
-
-const getSectionHeaderClasses = () => {
-  return 'px-3 py-2 text-xs font-medium text-muted-foreground border-l-2 bg-red-50 dark:bg-red-950/30 border-vusa-red'
-}
-
-const getItemClasses = () => {
-  return 'group mx-3 mb-4 p-5 rounded-lg border cursor-pointer hover:shadow-md transition-all duration-200 bg-white dark:bg-zinc-900 focus:outline-none focus:ring-2 focus:ring-vusa-red border-zinc-200 dark:border-zinc-700 hover:border-vusa-red/50 dark:hover:border-vusa-red/50 hover:bg-red-50/30 dark:hover:bg-red-950/10'
-}
-
-const getIconClasses = () => {
-  return 'w-6 h-6 rounded flex items-center justify-center bg-red-100 dark:bg-red-900/50 text-vusa-red dark:text-red-400'
-}
-
-const getLoadMoreClasses = () => {
-  return 'text-vusa-red hover:text-vusa-red-secondary hover:bg-red-50 dark:text-red-400 dark:hover:bg-red-950/30'
-}
-
-const formatDate = (dateValue: string | number | undefined) => {
-  if (!dateValue) return ''
-
-  try {
-    let date: Date
-
-    if (typeof dateValue === 'number') {
-      const timestamp = dateValue < 10000000000 ? dateValue * 1000 : dateValue
-      date = new Date(timestamp)
-    } else {
-      date = new Date(dateValue)
-    }
-
-    if (isNaN(date.getTime())) {
-      return String(dateValue)
-    }
-
-    return format(date, 'MMM dd, yyyy')
-  } catch {
-    return String(dateValue)
-  }
-}
-
-const stripHtml = (html: string): string => {
-  if (!html) return ''
-  const tmp = document.createElement('div')
-  tmp.innerHTML = html
-  return tmp.textContent || tmp.innerText || ''
-}
-
-const getItemDate = (item: any) => {
-  switch (props.type) {
-    case 'news': return item.publish_time
-    case 'documents': return item.document_date
-    case 'calendar': return item.date
-    default: return item.created_at
-  }
-}
-
-const getItemContent = (item: any) => {
-  let content = ''
-  switch (props.type) {
-    case 'news':
-      content = item.short || item.summary || item.content
-      break
-    case 'documents':
-      content = item.summary || item.content
-      break
-    case 'pages':
-      content = item.content || item.summary
-      break
-    case 'calendar':
-      content = item.description || item.summary
-      break
-    default:
-      content = item.summary || item.content || item.description
-  }
-
-  // Expand content length for better information display
-  if (content && content.length > 200) {
-    return content.slice(0, 200) + '...'
-  }
-  return content
-}
+// All utility functions now imported from useSearchUtils
 </script>
