+++ resources/js/Components/Public/Search/UnifiedResults.vue
@@ -16,7 +16,7 @@
         <div 
           v-for="(item, index) in unifiedResults" 
           :key="`${item.type}-${item.id}-${index}`"
-          :class="getUnifiedItemClasses(item)"
+          :class="getItemClasses()"
           @click="(event) => handleItemClick(item, event)"
           @keydown.enter="() => handleItemNavigation(item)"
           @keydown.space.prevent="() => handleItemNavigation(item)"
@@ -25,7 +25,7 @@
           :title="$t('search.click_to_open')"
           class="group cursor-pointer"
         >
-          <div class="flex items-start gap-4">
+          <div class="flex items-start gap-3">
             <!-- Content type indicator -->
             <div class="flex-shrink-0">
               <div :class="getIconClasses()">
@@ -49,14 +49,14 @@
               </div>
 
               <!-- Title -->
-              <h3 class="font-medium text-base leading-tight mb-2 group-hover:text-primary transition-colors">
+              <h3 class="font-medium text-sm leading-tight mb-1 group-hover:text-primary transition-colors">
                 <AisHighlight attribute="title" :hit="item" />
                 <AisHighlight v-if="!item.title && item.name" attribute="name" :hit="item" />
                 <span v-if="!item.title && !item.name">{{ item.name || item.title || $t('search.untitled') }}</span>
               </h3>
 
               <!-- Content preview -->
-              <div v-if="getItemContent(item)" class="text-sm text-muted-foreground leading-relaxed line-clamp-3 mb-1">
+              <div v-if="getItemContent(item)" class="text-xs text-muted-foreground leading-relaxed line-clamp-2">
                 {{ stripHtml(getItemContent(item)) }}
               </div>
             </div>
@@ -87,16 +87,10 @@ import { AisHighlight } from 'vue-instantsearch/vue3/es'
 import { Button } from '@/Components/ui/button'
 import { Badge } from '@/Components/ui/badge'
 import { trans as $t } from 'laravel-vue-i18n'
-import { format } from 'date-fns'
 import type { ContentType } from '@/Composables/useSearchController'
 import { useSearchService, type SearchService } from '@/Composables/useSearchService'
-
-// Import icons using the ~icons pattern like DutyForm.vue and SearchResultSection.vue
-import IconNews from '~icons/fluent/news24-regular'
-import IconPage from '~icons/fluent/document24-regular'
-import IconDocument from '~icons/fluent/document-checkmark24-regular'
-import IconCalendar from '~icons/fluent/calendar24-regular'
-import IconArrowUpRight from '~icons/fluent/arrow-up-right20-regular'
+import { useSearchUtils } from '@/Composables/useSearchUtils'
+import IconArrowRight from '~icons/fluent/arrow-right-16-filled'
 import IconSpinner from '~icons/fluent/spinner-ios20-filled'
 
 interface Props {
@@ -118,6 +112,18 @@ const emit = defineEmits<{
 // Use centralized search service - inject from parent or create new
 const searchService = inject<SearchService>('searchService') || useSearchService()
 const isLoadingMore = ref(false)
+
+// Use shared search utilities
+const { 
+  getIconComponent, 
+  formatDate, 
+  stripHtml, 
+  getItemDate, 
+  getItemContent,
+  getItemClasses,
+  getIconClasses,
+  getTypeBadgeClasses
+} = useSearchUtils()
 
 // Get unified results from search service
 const unifiedResults = computed(() => {
@@ -160,94 +166,7 @@ const handleItemNavigation = (item: any) => {
   emit('navigateToItem', item)
 }
 
-// Helper functions
-
-const getIconComponent = (type: string) => {
-  switch (type) {
-    case 'news': return IconNews
-    case 'pages': return IconPage
-    case 'documents': return IconDocument
-    case 'calendar': return IconCalendar
-    default: return IconPage
-  }
-}
-
-const getUnifiedItemClasses = (item: any) => {
-  return 'mx-3 mb-4 p-5 rounded-lg border cursor-pointer hover:shadow-md transition-all duration-200 bg-white dark:bg-zinc-900 focus:outline-none focus:ring-2 focus:ring-vusa-red focus:ring-offset-2 border-zinc-200 dark:border-zinc-700 hover:border-vusa-red/50 dark:hover:border-vusa-red/50 hover:bg-red-50/30 dark:hover:bg-red-950/10'
-}
-
-const getIconClasses = () => {
-  return 'w-6 h-6 rounded flex items-center justify-center bg-red-100 dark:bg-red-900/50 text-vusa-red dark:text-red-400'
-}
-
-const getTypeBadgeClasses = () => {
-  return 'text-xs border bg-red-100 text-vusa-red border-red-200 dark:bg-red-900/30 dark:text-red-300 dark:border-red-800'
-}
-
-const formatDate = (dateValue: string | number | undefined) => {
-  if (!dateValue) return ''
-  
-  try {
-    let date: Date
-    
-    if (typeof dateValue === 'number') {
-      const timestamp = dateValue < 10000000000 ? dateValue * 1000 : dateValue
-      date = new Date(timestamp)
-    } else {
-      date = new Date(dateValue)
-    }
-    
-    if (isNaN(date.getTime())) {
-      return String(dateValue)
-    }
-    
-    return format(date, 'MMM dd, yyyy')
-  } catch {
-    return String(dateValue)
-  }
-}
-
-const stripHtml = (html: string): string => {
-  if (!html) return ''
-  const tmp = document.createElement('div')
-  tmp.innerHTML = html
-  return tmp.textContent || tmp.innerText || ''
-}
-
-const getItemDate = (item: any) => {
-  switch (item.type) {
-    case 'news': return item.publish_time
-    case 'documents': return item.document_date
-    case 'calendar': return item.date
-    default: return item.created_at
-  }
-}
-
-const getItemContent = (item: any) => {
-  let content = ''
-  switch (item.type) {
-    case 'news': 
-      content = item.short || item.summary || item.content
-      break
-    case 'documents': 
-      content = item.summary || item.content
-      break
-    case 'pages': 
-      content = item.content || item.summary
-      break
-    case 'calendar': 
-      content = item.description || item.summary
-      break
-    default: 
-      content = item.summary || item.content || item.description
-  }
-  
-  // Expand content length for better information display
-  if (content && content.length > 200) {
-    return content.slice(0, 200) + '...'
-  }
-  return content
-}
+// All utility functions now imported from useSearchUtils
 
 // Check if more results are available from search service
 const hasMoreResults = computed(() => {
